🚚 Move: `lib/components/NewsSlider.svelte` -> `lib/components/news/NewsSlider.svelte` [#54]
