+++ src/lib/components/news/NewsSlider.svelte
